Stabilize P0-01 CI baseline: clippy clean, rustls provider, DHT teardown. Make workspace tests green on Windows/CI by fixing QUIC CryptoProvider, cancel-safe DHT re-announce, and router shutdown timeouts.

+++ crates/lp2ln-core-v2/src/node/topology_maintenance/session_reactions.rs
@@ -208,7 +208,7 @@ mod tests {
         }
 
         async fn send(&self, packet: crate::packet::Packet) -> Result<u64> {
-            let bytes = packet.wire_size_estimate() as u64;
+            let bytes = packet.wire_size_estimate();
             self.sent.lock().expect("sent lock").push(packet);
             Ok(bytes)
         }
